4.0.5 final release

+++ sync_repos.sh
@@ -1,13 +1,11 @@
 #!/bin/bash
 
 BILLINGREPO="https://github.com/OpenPEPPOL/peppol-bis-invoice-3.git"
-BILLINGBRANCH="2024-q4-release"
+BILLINGBRANCH="2025-Q2"
 LOGISTICSREPO="https://github.com/OpenPEPPOL/Logistics-bis.git"
-LOGISTICSBRANCH="2024-Q4"
+LOGISTICSBRANCH="2025-Q2-review"
 POACCREPO="https://github.com/OpenPEPPOL/poacc-upgrade-3.git"
-POACCBRANCH="2024-q4-release"
-
-
+POACCBRANCH="2025-Q2"
 
 # Function to clone and copy files
 sync_repo() {
@@ -18,12 +16,22 @@ sync_repo() {
 
   echo "Syncing $repo_url $source_path"
   git clone --depth=1 --branch "$branch" "$repo_url" temp-repo
-  # Ensure the target directory exists
-  mkdir -p "$target_path"
-  # Copy contents of the source directory to the target directory
-  cp -r temp-repo/"$source_path"/* "$target_path"
-  rm -rf temp-repo # Clean up temporary repo
+
+  if [ -d "temp-repo/$source_path" ]; then
+    mkdir -p "$target_path"
+    cp -rvf temp-repo/"$source_path"/* "$target_path"
+  elif [ -f "temp-repo/$source_path" ]; then
+    mkdir -p "$(dirname "$target_path")"
+    cp -vf temp-repo/"$source_path" "$target_path"
+  else
+    echo "Source path not found: $source_path"
+    rm -rf temp-repo
+    return 1
+  fi
+
+  rm -rf temp-repo
 }
+
 
 # Sync files from multiple repositories without './' in target path
 # Rules
@@ -36,17 +44,30 @@ sync_repo $LOGISTICSREPO $LOGISTICSBRANCH "rules/sch" "rules/sch"
 sync_repo $POACCREPO $POACCBRANCH "rules/examples/Cataloge wo response use cases" "rules/examples/Cataloge wo response use cases"
 sync_repo $POACCREPO $POACCBRANCH "rules/examples/Order use cases" "rules/examples/Order use cases"
 sync_repo $POACCREPO $POACCBRANCH "rules/examples/Order-response use cases" "rules/examples/Order-response use cases"
-sync_repo $LOGISTICSREPO $LOGISTICSBRANCH  "rules/examples" "rules/examples"
+# sync_repo $LOGISTICSREPO $LOGISTICSBRANCH "rules/examples" "rules/examples"
 
 # Code lists
 sync_repo $BILLINGREPO $BILLINGBRANCH "structure/codelist" "structure/codelist"
 sync_repo $LOGISTICSREPO $LOGISTICSBRANCH "structure/codelist" "structure/codelist"
 sync_repo $POACCREPO $POACCBRANCH "structure/codelist" "structure/codelist"
 
-# Code lists
-sync_repo $POACCREPO $POACCBRANCH "guides/profiles/1-catalogueonly" "guides/profiles/1-catalogueonly"
+# Tests
+sync_repo $BILLINGREPO $BILLINGBRANCH "rules/unit-UBL-SE" "rules/unit-UBL-SE" 
+sync_repo $LOGISTICSREPO $LOGISTICSBRANCH "rules/unit-common"  "rules/unit-common" 
+sync_repo $POACCREPO $POACCBRANCH "rules/unit-common"  "rules/unit-common" 
+sync_repo $POACCREPO $POACCBRANCH "rules/unit-common"  "rules/unit-order" 
+sync_repo $POACCREPO $POACCBRANCH "rules/unit-common"  "rules/order-agreement" 
+sync_repo $POACCREPO $POACCBRANCH "rules/unit-common"  "rules/order-response" 
+sync_repo $POACCREPO $POACCBRANCH "rules/unit-common"  "rules/unit-catalogue" 
+sync_repo $POACCREPO $POACCBRANCH "rules/unit-common"  "rules/unit-invoice-response" 
+
+# Guides
+sync_repo $POACCREPO $POACCBRANCH "guides/profiles/28-ordering" "guides/profiles/28-ordering"
+sync_repo $POACCREPO $POACCBRANCH "guides/profiles/3-order-only" "guides/profiles/3-order-only"
+sync_repo $POACCREPO $POACCBRANCH "guides/profiles/42-orderagreement" "guides/profiles/42-orderagreement"
+sync_repo $POACCREPO $POACCBRANCH "guides/profiles/63-invoiceresponse" "guides/profiles/63-invoiceresponse"
 sync_repo $POACCREPO $POACCBRANCH "guides/profiles/64-catalogue-wo-response" "guides/profiles/64-catalogue-wo-response"
-#sync_repo $LOGISTICSREPO $LOGISTICSBRANCH "guides/profiles/66-advanceddespatchadvice" "guides/profiles/66-advanceddespatchadvice"
+# sync_repo $LOGISTICSREPO $LOGISTICSBRANCH "guides/profiles/66-advanceddespatchadvice" "guides/profiles/66-advanceddespatchadvice"
 sync_repo $LOGISTICSREPO $LOGISTICSBRANCH "guides/profiles/67-advanceddespatchadvice_w_receiptadvice" "guides/profiles/67-advanceddespatchadvice_w_receiptadvice"
 sync_repo $LOGISTICSREPO $LOGISTICSBRANCH "guides/profiles/68-weightstatement" "guides/profiles/68-weightstatement"
 sync_repo $LOGISTICSREPO $LOGISTICSBRANCH "guides/profiles/69-transportexecutionplan_w_request" "guides/profiles/69-transportexecutionplan_w_request"
@@ -54,3 +75,4 @@ sync_repo $LOGISTICSREPO $LOGISTICSBRANCH "guides/profiles/70-transportexecution
 sync_repo $LOGISTICSREPO $LOGISTICSBRANCH "guides/profiles/71-waybill" "guides/profiles/71-waybill"
 sync_repo $LOGISTICSREPO $LOGISTICSBRANCH "guides/profiles/72-transportationstatus_w_request" "guides/profiles/72-transportationstatus_w_request"
 sync_repo $LOGISTICSREPO $LOGISTICSBRANCH "guides/profiles/73-transportationstatus" "guides/profiles/73-transportationstatus"
+sync_repo $LOGISTICSREPO $LOGISTICSBRANCH "guides/shared" "guides/shared"
